extrapolated data from homepage to util/data
