+++ frontend/src/utils/helper/OldData.js
@@ -1,0 +1,65 @@
+{
+  /* <Chart data={data}>
+                <Axis primary type="ordinal" position="left" />
+                <Axis type="linear" stacked position="bottom" />
+                <Series type={Bar} />
+                <Cursor primary />
+                <Cursor />
+                <Tooltip />
+            </Chart> */
+}
+
+{
+  /* <div className="homepage-license">
+                <List className="license-type1">
+                  <ListItem>SEL Hours : {this.state.sel}</ListItem>
+                  <ListItem className="LicenseType">
+                    SES Hours : {this.state.ses}
+                  </ListItem>
+                </List>
+                <List className="license-type2">
+                  <ListItem className="LicenseType1">
+                    MEL Hours : {this.state.mel}
+                  </ListItem>
+                  <ListItem className="LicenseType">
+                    MES Hours : {this.state.mes}
+                  </ListItem>
+                </List>
+              </div>
+              <CardContent />
+              <CardContent className="TotalsBodyChildren">
+                <div classname="totals-1">
+                  <p>Cross Country : {this.state.cross_country}</p>
+                  <p>No. Instr. App. : {this.state.no_instument_app}</p>
+                  <p>No. Ldg. : {this.state.no_ldg}</p>
+                  <p>Day : {this.state.day}</p>
+                  <p>Night : {this.state.night}</p>
+                </div>
+                <div classname="totals-2">
+                  <p>Actual : {this.state.actual_instr}</p>
+                  <p>SIM : {this.state.sim_instr}</p>
+                  <p>Grnd Trnr. :</p>
+                  <p>PIC : {this.state.pic}</p>
+                  <p>Dual Rec. : {this.state.dual_rec}</p>
+                </div>
+              </CardContent> */
+}
+
+// var stackedBar = new Chart({
+//   type: "bar",
+//   data: data,
+//   options: {
+//     scales: {
+//       xAxes: [
+//         {
+//           stacked: true
+//         }
+//       ],
+//       yAxes: [
+//         {
+//           stacked: true
+//         }
+//       ]
+//     }
+//   }
+// });
